refactor : (S3) DDD layer에 맞는 디렉토리 설정
